% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootstrap.costs}
\alias{bootstrap.costs}
\title{Boootstrap samples of DAP/MAP costs by supplier, region and scenario}
\format{
An object of class \code{data.frame} with 18000 rows and 8 columns.
}
\usage{
bootstrap.costs
}
\description{
This compilation provides bootstrap samples for suppliers' trade costs in the distributed DAP/MAP market model.
The compilation covers 5 suppliers, 9 regions and FAO "Business as Usual" and "Stratified Societies" scenarios for crop production in year 2030.
Cost values are generated by supplier, region and scenario using regression model with wild unrestricted residual bootstrap (Davidson and MacKinnon 2010).
}
\details{
Values are measured on the [0,1] price/cost scale throughout this compilation.

The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

This compilation covers the following phosphate companies and groups of companies:
\itemize{
  \item Chinese exporters
  \item Maaden https://www.maaden.com.sa/
  \item Mosaic http://www.mosaicco.com/
  \item PhosAgro/EuroChem https://phosagro.com/, https://www.eurochemgroup.com/
  \item OCP https://www.ocpgroup.ma/
}

Regional classification includes: Africa, East Asia, Eastern Europe and Central Asia, Latin America, North America, Oceania, South Asia, Western and Central Europe, and Western Asia.

The data on DAP/MAP fertilizer trade volumes in years 2013-2017 in \code{dapmap.markets} package is used for estimation. \code{bootstrap.demand} is used for projection.

The data processing of original values into a sample includes the following steps:
\enumerate{
  \item Filter original data to exclude flows smaller than 1\% of total import flows to a region in the year.
Add excluded volume to regional domestic supply. Total consumption by region and year in corrected data will be equal to total consumption in the raw data.
  \item Calculate nominal market share in global demand based on the filtered data (by year and region). 
  \item Find growth rate of Latin America market share in global demand relative to the share in year 2014 (by year).
  \item Find real market share in global demand by dividing the nominal value by the growth rate of Latin America market share (by year and region).
  \item Calculate trade cost from the equilibrium volumes for a supplier present on a market with the assumption that trade cost for Latin America market equals 0 (by supplier, region and year).
  \item Compose a sample of differences in real market shares relative to base year 2014, 
and a sample of differences in trade costs relative to base year 2014 (for years 2015-2017).
}

The bootstrap data-generating process includes the following steps:
\enumerate{
  \item Estimate regression coefficients and residuals using least squares on the joint sample: regress differences in trade costs on differences in real market shares.
  \item Resample with replacement from centered residuals using wild unrestricted residual bootstrap with Rademacher distribution.
  \item Calculate forecast differences in real market shares for year 2030 relative to base year 2014 (by scenario).
  \item Estimate differences in trade costs in 2030 as predicted values from the model with forecast differences in real market shares for year 2030 from step 3, 
estimated coefficients from step 1 and new bootstrap residuals from step 2.
  \item Calculate trade costs in 2030 using the estimated differences from step 4 and reference trade costs in base year 2014. 
  \item Repeat steps 2-5 1000 times and collect values in the bootstrap samples of DAP/MAP costs by supplier, region and scenario.
}

In total, there are 101 observations in the joint sample of differences for regression estimation.
}
\references{
\itemize{
  \item Davidson, R., & MacKinnon, J. (2010). Wild Bootstrap for IV Regression. Journal of Business & Economic Statististics, 28(1):128-144.
}
}
\keyword{datasets}
