% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootstrap.demand}
\alias{bootstrap.demand}
\title{Boootstrap samples of DAP/MAP demand by region and scenario}
\format{
An object of class \code{data.frame} with 2000 rows and 10 columns.
}
\usage{
bootstrap.demand
}
\description{
This compilation provides bootstrap samples for regional demand in the distributed DAP/MAP market model.
The compilation covers 9 regions and FAO "Business as Usual" and "Stratified Societies" scenarios for crop production in year 2030.
Demand values are generated by region and scenario using instrumental variable regression model with wild unrestricted residual bootstrap (Davidson and MacKinnon 2010).
}
\details{
Metric kilotonnes of P2O5 per annum are used for volumes throughout this compilation.

The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

Regional classification includes: Africa, East Asia, Eastern Europe and Central Asia, Latin America, North America, Oceania, South Asia, Western and Central Europe, and Western Asia.

Regression variables include:
\itemize{
  \item z: phosphorus use in selected crops production (instrument),
  \item x: total phosphorus use in agriculture (endogenous variable in the reduced-form equation),
  \item y: DAP/MAP consumption (endogenous variable in the structural equation).
}

The data in \code{dapmap.markets} and \code{puse} packages is used for estimation.
Original values for years 2007-2017 are smoothed using central moving average with span 3.

The bootstrap data-generating process includes the following steps:
\enumerate{
  \item Estimate original regression coefficients and residuals using two-stage least squares and smoothed data for years 2007-2017 (9 observations): regress x on z, regress y on prediced x.
  \item Resample with replacement disturbance terms in reduced-form and structural equations using wild unrestricted bootstrap with Rademacher distribution on centered and corrected residuals.
  \item Compile new bootstrap sample from equations with instrument, bootstrap residuals and estimated coefficients.
  \item Estimate bootstrap coefficients and residuals using two-stage least squares and new bootstrap sample. 
  \item Generate new disturbance terms as in step 2.
  \item Estimate DAP/MAP consumption in 2030 as a predicted value from the model with bootstrap coefficients from step 4 and new bootstrap residuals from step 5.
  \item Repeat steps 2-6 1000 times and collect values in the bootstrap sample of DAP/MAP demand.
}

In total, the bootstrap data-generating process for DAP/MAP demand is run 18 times to obtain samples for each region and scenario.
}
\references{
\itemize{
  \item Davidson, R., & MacKinnon, J. (2010). Wild Bootstrap for IV Regression. Journal of Business & Economic Statististics, 28(1):128-144.
}
}
\keyword{datasets}
