% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootstrap.supply}
\alias{bootstrap.supply}
\title{Boootstrap samples of DAP/MAP supply by supplier and scenario}
\format{
An object of class \code{data.frame} with 2000 rows and 6 columns.
}
\usage{
bootstrap.supply
}
\description{
This compilation provides bootstrap samples for producers' supply in the distributed DAP/MAP market model.
The compilation covers 5 suppliers and FAO "Business as Usual" and "Stratified Societies" scenarios for crop production in year 2030.
Supply values are generated by supplier and scenario using wild bootstrap (Davidson and MacKinnon 2010).
It is assumed that international suppliers aim to maintain their shares in global market supply in 2030 by adding capacity as necessary ("Commodity Classic" scenario in IFA 2018).
}
\details{
Metric kilotonnes of P2O5 per annum are used for volumes throughout this compilation.

The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

This compilation covers the following phosphate companies and groups of companies:
\itemize{
  \item Chinese exporters
  \item Maaden https://www.maaden.com.sa/
  \item Mosaic http://www.mosaicco.com/
  \item PhosAgro/EuroChem https://phosagro.com/, https://www.eurochemgroup.com/
  \item OCP https://www.ocpgroup.ma/
}

The data in \code{dapmap.markets} package for years 2013-2017 is used for estimation. \code{bootstrap.demand} is used for projection.

The bootstrap data-generating process includes the following steps:
\enumerate{
  \item Estimate mean market share and deviations from the mean using data for years 2007-2017 by supplier (25 observations in total).
  \item Resample with replacement from the joint sample of centered deviations from the mean using wild bootstrap with Rademacher distribution .
  \item Compile new bootstrap sample of market shares from observed means and bootstrap deviations from the mean.
  \item Estimate DAP/MAP supply from bootstrap market shares and world demand in \code{bootstrap.demand}.
  \item Repeat steps 2-4 1000 times and collect values in the bootstrap sample of DAP/MAP supply.
}

In total, the bootstrap data-generating process for DAP/MAP supply is run 2 times to obtain samples for each scenario.
}
\references{
\itemize{
  \item Davidson, R., & MacKinnon, J. (2010). Wild Bootstrap for IV Regression. Journal of Business & Economic Statististics, 28(1):128-144.
  \item IFA (2018). IFA 2030 Scenarios. Downloaded from https://www.fertilizer.org/Public/About_IFA/IFA2030.aspx
}
}
\keyword{datasets}
