% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{dapmap.parameters}
\alias{dapmap.parameters}
\title{Bootstrap parameters for distributed DAP/MAP market model}
\seealso{
\code{\link{bootstrap.demand}}, \code{\link{bootstrap.supply}}, \code{\link{bootstrap.costs}}
}
